+++ src/components/Stack.js
@@ -1,6 +1,6 @@
 import React, {useState} from 'react';
 import { useSpring, animated } from 'react-spring';
-import StackForm from './Stack-Form';
+import StackForm from './Forms/Stack-Form';
 import Container from 'react-bootstrap/Container';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
